refactor(higgsfield): decompose commands/video/common/image into focused modules
Extract 12 new focused modules from the 4 remaining higgsfield files
that were flagged by qlty file-complexity:

From higgsfield-common.mjs (283 → ~50):
  - higgsfield-browser.mjs: browser launch, navigation, modal dismissal
  - higgsfield-output.mjs: output dirs, sidecars, dedup, image downloads
  - higgsfield-discovery.mjs: site discovery, login
  - higgsfield-batch-infra.mjs: batch operations infrastructure

From higgsfield-commands.mjs (334 → ~50):
  - higgsfield-asset-chain.mjs: asset chain operations
  - higgsfield-pipeline.mjs: video production pipeline
  - higgsfield-tests.mjs: auth health check, smoke test, self-tests
  - higgsfield-studio-commands.mjs: studio/tool commands

From higgsfield-video.mjs (297 → ~80):
  - higgsfield-lipsync.mjs: lipsync generation
  - higgsfield-batch-video.mjs: batch video/lipsync operations
  - higgsfield-video-download.mjs: video download helpers

From higgsfield-image.mjs (102 → ~50):
  - higgsfield-image-helpers.mjs: generation detection/polling

Original files retain backward-compatible re-exports.
Total qlty smells reduced from 92 to 27 (-65).

Resolves #19223
